Semana7/microblog/microblog/requirements.txt
django
djangorestframework
django-ratelimit
django-cors-headers